--- a/tooling/build/scripts/publisher.sh
+++ b/tooling/build/scripts/publisher.sh
@@ -16,22 +16,6 @@ if [ -z "$ISOMER_BUILD_REPO_BRANCH" ]; then
   ISOMER_BUILD_REPO_BRANCH=$(curl https://api.github.com/repos/opengovsg/isomer/releases/latest | jq -r '.tag_name')
 fi
 
-# Define S3 cache path
-NODE_MODULES_CACHE_PATH="s3://$S3_CACHE_BUCKET_NAME/cache/$ISOMER_BUILD_REPO_BRANCH/node_modules.tar.gz"
-
-# Try to fetch cached node_modules from S3
-echo "Fetching cached node_modules..."
-start_time=$(date +%s)
-aws s3 cp $NODE_MODULES_CACHE_PATH node_modules.tar.gz || true
-if [ -f "node_modules.tar.gz" ]; then
-  tar -xzf node_modules.tar.gz
-  rm node_modules.tar.gz
-  echo "Using cached node_modules"
-else
-  echo "No cached node_modules found"
-fi
-calculate_duration $start_time
-
 # Cloning the repository
 echo "Cloning repository..."
 start_time=$(date +%s)
@@ -53,24 +37,32 @@ echo $(git branch)
 # Perform a clean of npm cache
 npm cache clean --force
 
-# Install dependencies if cache miss or update dependencies if cache hit
-if [ ! -d "node_modules" ]; then
+# Try to fetch cached node_modules from S3
+echo "Fetching cached node_modules..."
+NODE_MODULES_CACHE_PATH="s3://$S3_CACHE_BUCKET_NAME/cache/$ISOMER_BUILD_REPO_BRANCH/node_modules.tar.gz"
+aws s3 cp $NODE_MODULES_CACHE_PATH node_modules.tar.gz || true
+if [ -f "node_modules.tar.gz" ]; then
+  echo "node_modules.tar.gz found in cache"
+  
+  echo "Using cached node_modules"
+  start_time=$(date +%s)
+  tar -xzf node_modules.tar.gz
+  rm node_modules.tar.gz
+  calculate_duration $start_time
+else
+  echo "node_modules.tar.gz not found in cache"
+
   echo "Installing dependencies..."
   start_time=$(date +%s)
   npm ci
   calculate_duration $start_time
 
-  # Cache node_modules to S3 only if we had to install them
   echo "Caching node_modules..."
   start_time=$(date +%s)
   tar -czf node_modules.tar.gz node_modules/
   aws s3 cp node_modules.tar.gz $NODE_MODULES_CACHE_PATH
   rm node_modules.tar.gz
   echo "Cached node_modules"
-  calculate_duration $start_time
-else
-  start_time=$(date +%s)
-  echo "node_modules already exists, skipping npm ci"
   calculate_duration $start_time
 fi