--- a/tooling/build/scripts/publisher.sh
+++ b/tooling/build/scripts/publisher.sh
@@ -39,7 +39,7 @@ npm cache clean --force
 
 # Try to fetch cached node_modules from S3
 echo "Fetching cached node_modules..."
-NODE_MODULES_CACHE_PATH="s3://$S3_CACHE_BUCKET_NAME/cache/$ISOMER_BUILD_REPO_BRANCH/node_modules.tar.gz"
+NODE_MODULES_CACHE_PATH="s3://$S3_CACHE_BUCKET_NAME/$ISOMER_BUILD_REPO_BRANCH/isomer/node_modules.tar.gz"
 aws s3 cp $NODE_MODULES_CACHE_PATH node_modules.tar.gz || true
 if [ -f "node_modules.tar.gz" ]; then
   echo "node_modules.tar.gz found in cache"
@@ -88,7 +88,6 @@ calculate_duration $start_time
 
 # Build site
 echo "Building site..."
-start_time=$(date +%s)
 rm -rf ../../../template/schema
 rm -rf ../../../template/data
 mv schema/ ../../../template/
@@ -97,8 +96,33 @@ cp sitemap.json ../../../template/public/
 mv sitemap.json ../../../template/
 cd ../../../template
 echo $(pwd)
-npm ci
-calculate_duration $start_time
+echo "Fetching cached tooling-template node_modules..."
+TOOLING_TEMPLATE_NODE_MODULES_CACHE_PATH="s3://$S3_CACHE_BUCKET_NAME/$ISOMER_BUILD_REPO_BRANCH/isomer-tooling-template/node_modules.tar.gz"
+aws s3 cp $TOOLING_TEMPLATE_NODE_MODULES_CACHE_PATH node_modules.tar.gz || true
+if [ -f "node_modules.tar.gz" ]; then
+  echo "node_modules.tar.gz found in cache"
+  
+  echo "Using cached node_modules"
+  start_time=$(date +%s)
+  tar -xzf node_modules.tar.gz
+  rm node_modules.tar.gz
+  calculate_duration $start_time
+else
+  echo "node_modules.tar.gz not found in cache"
+
+  echo "Installing dependencies..."
+  start_time=$(date +%s)
+  npm ci
+  calculate_duration $start_time
+
+  echo "Caching node_modules..."
+  start_time=$(date +%s)
+  tar -czf node_modules.tar.gz node_modules/
+  aws s3 cp node_modules.tar.gz $TOOLING_TEMPLATE_NODE_MODULES_CACHE_PATH
+  rm node_modules.tar.gz
+  echo "Cached node_modules"
+  calculate_duration $start_time
+fi
 
 # Prebuild
 echo "Prebuilding..."