## Problem

<!-- What problem are you trying to solve? What issue does this close? -->

Reduce build time and cost by caching

partially closes https://linear.app/ogp/issue/ISOM-1661/optimise-site-build-times

## Solution

<!-- How did you solve the problem? -->

**Breaking Changes**

<!-- Does this PR contain any backward incompatible changes? If so, what are they and should there be special considerations for release? -->

- [ ] Yes - this PR contains breaking changes
  - Details ...
- [x] No - this PR is backwards compatible

**Improvements**:

TL;DR - cache `node_modules` and save ~2+ mins per build

approach
- storing cached files (based on release tag version) in S3 as there's no guarantee local cache will persist between builds given the low frequency of builds and the on-demand setup of our codebuild instances
- since releases are not very often (~1-2 times weekly), this reduces build time for every build across sites
- e.g. MOH-corp now build at ~14 mins when testing with PROD data

how:
- cache `node_modules` (isomer)
  - before: between 80 to 120 seconds to install
  - after: ~6 seconds to fetch from S3
  - if cache miss, takes ~6 seconds (up to 40 seconds) to upload to S3 too after installing
- cache all the `node_modules` (tooling/template)
  - before: ~50 (9 + 33 + 8) seconds to install
  - after: ~0 seconds to fetch from S3
  - if cache miss, takes ~1 seconds (up to 17seconds)to upload to S3 too after installing

note
- did not cache `.next/cache` because it is too big and does not really speed up `next build` much. In fact, it's actually slower if we have to download the cache from S3 than actually building it on the spot
- while ideally we use codebuild in-built cache mechanism in buildspec, but:
  - it's easier to maintain everything in the publisher script 
  - cache are tagged to dynamic keys like `ISOMER_BUILD_REPO_BRANCH` which are fetch on build-time, which can't be done easily since we are using pulumi to define the stack