@adriangohjw 2024-11-18
@dcshzj can help review https://github.com/opengovsg/isomer-next-infra/pull/27 too? this can't be merged before that 😬